UserApi refactor

+++ db/src/main/java/museum/mapper/AuthorMapper.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AuthorMapper extends CrudRepository<AuthorModel, Long> {
+public interface AuthorMapper extends CrudRepository<AuthorModel, Integer> {
   AuthorModel findByFullName(String fullName );
   //List<AuthorModel> find( );
